LayoutTests/accessibility/global-aria-attributes-invalidate-presentational.html

<!DOCTYPE HTML PUBLIC "-//IETF//DTD HTML//EN">
<html>
<head>
<script src="../fast/js/resources/js-test-pre.js"></script>
</head>
<body id="body">

<img id="item1" src="foo.jpg" alt="test1" role="presentation" aria-atomic="true">
<img id="item2" src="foo.jpg" alt="test1" role="presentation" aria-busy="true">
<img id="item3" src="foo.jpg" alt="test1" role="presentation" aria-controls="item1">
<img id="item4" src="foo.jpg" alt="test1" role="presentation" aria-describedby="item1">
<img id="item5" src="foo.jpg" alt="test1" role="presentation" aria-disabled="true">
<img id="item6" src="foo.jpg" alt="test1" role="presentation" aria-dropeffect="true">
<img id="item7" src="foo.jpg" alt="test1" role="presentation" aria-haspopup="true">
<img id="item8" src="foo.jpg" alt="test1" role="presentation" aria-flowto="item1">
<img id="item9" src="foo.jpg" alt="test1" role="presentation" aria-invalid="true">
<img id="item10" src="foo.jpg" alt="test1" role="presentation" aria-label="test">
<img id="item11" src="foo.jpg" alt="test1" role="presentation" aria-labelledby="item1">
<img id="item12" src="foo.jpg" alt="test1" role="presentation" aria-live="polite">
<img id="item13" src="foo.jpg" alt="test1" role="presentation" aria-owns="item1">
<img id="item14" src="foo.jpg" alt="test1" role="presentation" aria-relevant="all">

<p id="description"></p>
<div id="console"></div>

<script>

    description("This tests that the presence of any global ARIA attribute invalidates a presentational role.");

    if (window.accessibilityController) {
        for (var k = 1; k < 15; k++) {
            var item = accessibilityController.accessibleElementById("item" + k);
            shouldBe("item.role", "'AXRole: AXImage'");
        }
    }

</script>

<script src="../fast/js/resources/js-test-post.js"></script>
</body>
</html>
