LayoutTests/accessibility/global-aria-attributes-invalidate-presentational-expected.txt


This tests that the presence of any global ARIA attribute invalidates a presentational role.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS item.role is 'AXRole: AXImage'
PASS item.role is 'AXRole: AXImage'
PASS item.role is 'AXRole: AXImage'
PASS item.role is 'AXRole: AXImage'
PASS item.role is 'AXRole: AXImage'
PASS item.role is 'AXRole: AXImage'
PASS item.role is 'AXRole: AXImage'
PASS item.role is 'AXRole: AXImage'
PASS item.role is 'AXRole: AXImage'
PASS item.role is 'AXRole: AXImage'
PASS item.role is 'AXRole: AXImage'
PASS item.role is 'AXRole: AXImage'
PASS item.role is 'AXRole: AXImage'
PASS item.role is 'AXRole: AXImage'
PASS successfullyParsed is true

TEST COMPLETE
